remove code from actions into saga file
