+++ src/reducers/index.js
@@ -1,7 +1,8 @@
-import searchAndFilter from './searchAndFilter.reducer';
 import { combineReducers } from 'redux';
 import builder from './builder.reducer';
-import devicesList from './devicesList.reducer';
+import {
+  devicesList,
+  searchAndFilter } from './devicesList.reducer';
 import { loadUsersReducer } from './users.reducer';
 import notificationsReducer from './notifications.reducer';
