+++ src/sagas/index.js
@@ -1,25 +1,11 @@
 import { all, takeEvery } from 'redux-saga/effects';
-import {
-  loadDevicesSaga,
-  loadDevices,
-  loadDevice,
-  deleteDevice
-} from '../actions/devices.action';
 import currentUsersSaga from './users.saga';
 import notificationsSaga from './notifications.saga';
 import { watchAddDevice } from './builder.saga';
-
-export function* watchLoadDevices () {
-  yield takeEvery('LOAD_DEVICES', loadDevicesSaga);
-}
-
-export function* watchLoadDeviceAsync () {
-  yield takeEvery('LOAD_DEVICE_ASYNC', loadDevice);
-}
-
-export function* watchDeleteDeviceAsync () {
-  yield takeEvery('DELETE_DEVICE_ASYNC', deleteDevice);
-}
+import {
+  watchLoadDevices,
+  watchLoadDeviceAsync,
+  watchDeleteDeviceAsync } from './devices.saga.js';
 
 export default function* rootSaga () {
   yield all([
